+++ src/main/java/com/ddobang/backend/domain/diary/repository/DiaryRepositoryImpl.java
@@ -5,8 +5,8 @@ import static com.ddobang.backend.domain.diary.entity.QDiaryStat.*;
 import static com.ddobang.backend.domain.region.entity.QRegion.*;
 import static com.ddobang.backend.domain.store.entity.QStore.*;
 import static com.ddobang.backend.domain.theme.entity.QTheme.*;
-import static com.ddobang.backend.domain.theme.entity.QThemeTag.*;
 import static com.ddobang.backend.domain.theme.entity.QThemeTagMapping.*;
+import static com.ddobang.backend.domain.theme.tag.entity.QThemeTag.*;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,8 +17,8 @@ import org.springframework.stereotype.Repository;
 import com.ddobang.backend.domain.diary.dto.request.DiaryFilterRequest;
 import com.ddobang.backend.domain.diary.entity.Diary;
 import com.ddobang.backend.domain.member.entity.Member;
-import com.ddobang.backend.domain.theme.entity.QThemeTag;
 import com.ddobang.backend.domain.theme.entity.QThemeTagMapping;
+import com.ddobang.backend.domain.theme.tag.entity.QThemeTag;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Order;
 import com.querydsl.core.types.OrderSpecifier;
