+++ src/main/java/com/ddobang/backend/domain/diary/repository/DiaryStatRepositoryImpl.java
@@ -3,8 +3,8 @@ package com.ddobang.backend.domain.diary.repository;
 import static com.ddobang.backend.domain.diary.entity.QDiaryStat.*;
 import static com.ddobang.backend.domain.theme.entity.QTheme.*;
 import static com.ddobang.backend.domain.theme.entity.QThemeStat.*;
-import static com.ddobang.backend.domain.theme.entity.QThemeTag.*;
 import static com.ddobang.backend.domain.theme.entity.QThemeTagMapping.*;
+import static com.ddobang.backend.domain.theme.tag.entity.QThemeTag.*;
 
 import java.util.LinkedHashMap;
 import java.util.List;
