+++ src/main/java/com/ddobang/backend/domain/party/repository/PartyRepositoryImpl.java
@@ -21,9 +21,9 @@ import com.ddobang.backend.domain.party.types.PartyMemberStatus;
 import com.ddobang.backend.domain.party.types.PartyStatus;
 import com.ddobang.backend.domain.store.entity.QStore;
 import com.ddobang.backend.domain.theme.entity.QTheme;
-import com.ddobang.backend.domain.theme.entity.QThemeTag;
 import com.ddobang.backend.domain.theme.entity.QThemeTagMapping;
 import com.ddobang.backend.domain.theme.entity.Theme;
+import com.ddobang.backend.domain.theme.tag.entity.QThemeTag;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
