+++ src/main/java/com/ddobang/backend/domain/theme/entity/Theme.java
@@ -7,6 +7,7 @@ import org.hibernate.validator.constraints.Length;
 
 import com.ddobang.backend.domain.store.entity.Store;
 import com.ddobang.backend.domain.theme.dto.request.ThemeForAdminRequest;
+import com.ddobang.backend.domain.theme.tag.entity.ThemeTag;
 import com.ddobang.backend.global.entity.BaseTime;
 
 import jakarta.persistence.CascadeType;
