+++ src/main/java/com/ddobang/backend/domain/theme/tag/repository/ThemeTagRepository.java
@@ -1,4 +1,4 @@
-package com.ddobang.backend.domain.theme.repository;
+package com.ddobang.backend.domain.theme.tag.repository;
 
 import java.util.Collection;
 import java.util.List;
@@ -8,8 +8,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import com.ddobang.backend.domain.theme.dto.response.ThemeTagResponse;
-import com.ddobang.backend.domain.theme.entity.ThemeTag;
+import com.ddobang.backend.domain.theme.tag.dto.ThemeTagResponse;
+import com.ddobang.backend.domain.theme.tag.entity.ThemeTag;
 
 /**
  * ThemeTagRepository
